Open stdout and stderr as text streams
Nmap stdout and sterr is being used as str. This opens the streams
as text, so read()/readline() returns a str instead of a byte, avoiding
the need to manually decode from str to byte on each read().

This fixes savon-noir/python-libnmap#63 which is an endless loop
caused by enumerate() waiting for a '' as sentinel but receiving a b''.

Also, NmapProcess.__process_event() expects stderr as str, using a byte
breaks it, which in turn breaks NmapProcess.run_background() and
NmapProcess.sudo_run_background()