Fix CVE-2022-30284 by validating targets
This change adds an extra validation to `NmapProcess` that does
some basic checks on the `targets` parameter.

It checks that no unexpected characters are present in each target,
and that if dashes are present, they are not located at the start
or end of the string (a valid FQDN can have dashes except in those
positions).

This last check fixes CVE-2022-30284, which allowed to inject
extra command line options by using a target such as
`127.0.0.1,--script,http-fetch....`.